int -> long int

+++ src/utility/GridIterators/wrappers_for_templates.h
@@ -4,19 +4,19 @@
 /* headers for functions called in wrappers_for_templates.c */
       
 /* header for GMRES in templates */
-int gmres_(int *N, double *B, double *X, int *RESTRT, 
-            double *WORK, int *LDW, double *H, int *LDH,
-            int *ITER, double *RESID,
+int gmres_(long int *N, double *B, double *X, long int *RESTRT, 
+            double *WORK, long int *LDW, double *H, long int *LDH,
+            long int *ITER, double *RESID,
             int (*matvec)(double *alpha, double *x, double *beta, double *y),
             int (*psolve)(double *x, double *b),
-            int *INFO);
-int bicgstab_(int *N, double *B, double *X, double *WORK, int *LDW, 
-              int *ITER, double *RESID,
+            long int *INFO);
+int bicgstab_(long int *N, double *B, double *X, double *WORK, long int *LDW, 
+              long int *ITER, double *RESID,
               int (*matvec)(double *alpha, double *x, double *beta, double *y),
               int (*psolve)(double *x, double *b),
-              int *INFO);
-int cgs_(int *N, double *B, double *X, double *WORK, int *LDW, 
-              int *ITER, double *RESID,
+              long int *INFO);
+int cgs_(long int *N, double *B, double *X, double *WORK, long int *LDW, 
+              long int *ITER, double *RESID,
               int (*matvec)(double *alpha, double *x, double *beta, double *y),
               int (*psolve)(double *x, double *b),
-              int *INFO);
+              long int *INFO);
